...Done?
Lots going on in this one. I changed up the formatting of the new files a bit so that they would be more compatable (made the #RMID column consistant so that it could be used as the index when joining up the lines (see lines 52-62 in the code)). Plotted the new data and used correct errors (including different positive and negative error values for the new files).